RAC-157 refactor: salary application 분리
