+++ src/main/java/com/postgraduate/domain/senior/application/usecase/SeniorMyPageUseCase.java
@@ -1,10 +1,5 @@
 package com.postgraduate.domain.senior.application.usecase;
 
-import com.postgraduate.domain.salary.application.dto.res.SalaryDetailResponse;
-import com.postgraduate.domain.salary.application.dto.res.SalaryInfoResponse;
-import com.postgraduate.domain.salary.application.mapper.SalaryMapper;
-import com.postgraduate.domain.salary.domain.entity.Salary;
-import com.postgraduate.domain.salary.domain.service.SalaryGetService;
 import com.postgraduate.domain.senior.application.dto.req.SeniorMyPageProfileRequest;
 import com.postgraduate.domain.senior.application.dto.res.SeniorInfoResponse;
 import com.postgraduate.domain.senior.application.mapper.SeniorMapper;
@@ -19,8 +14,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
-import java.util.List;
 import java.util.Optional;
 
 import static java.util.Optional.ofNullable;
@@ -29,9 +22,7 @@ import static java.util.Optional.ofNullable;
 @RequiredArgsConstructor
 @Transactional
 public class SeniorMyPageUseCase {
-    private static final int SALARY_DATE = 10;
     private final SeniorGetService seniorGetService;
-    private final SalaryGetService salaryGetService;
     private final SeniorUpdateService seniorUpdateService;
     private final UserUpdateService userUpdateService;
 
@@ -46,33 +37,5 @@ public class SeniorMyPageUseCase {
         userUpdateService.updateSeniorMyPage(user.getUserId(), myPageProfileRequest);
         Senior senior = seniorGetService.byUser(user);
         seniorUpdateService.updateMyPageProfile(senior, myPageProfileRequest);
-    }
-
-    public SalaryInfoResponse getSalary(User user) {
-        Senior senior = seniorGetService.byUser(user);
-        LocalDate settlementDate = getSettlementDate();
-        List<Salary> salaries = salaryGetService.bySeniorAndSalaryDate(senior, settlementDate);
-        int pay = getAmount(salaries);
-        return new SalaryInfoResponse(settlementDate, pay); //TODO 수수료
-    }
-
-    private LocalDate getSettlementDate() {
-        LocalDate now = LocalDate.now();
-        return now.getDayOfMonth() <= SALARY_DATE
-                ? now.withDayOfMonth(SALARY_DATE)
-                : now.plusMonths(1).withDayOfMonth(SALARY_DATE);
-    }
-
-    private int getAmount(List<Salary> salaries) {
-        return salaries.stream()
-                .map(salary -> salary.getMentoring().getPay())
-                .mapToInt(Integer::intValue)
-                .sum();
-    }
-
-    public List<SalaryDetailResponse> getSalaryDetail(User user, Boolean status) {
-        Senior senior = seniorGetService.byUser(user);
-        List<Salary> salaries = salaryGetService.bySeniorAndStatus(senior, status);
-        return salaries.stream().map(SalaryMapper::mapToSalaryDetail).toList();
     }
 }
